Allow to also play specified playlists and tracks
Every item on the command line is a so called track group. Reordering does not affect tracks across track group boundaries.
The arguments genre and time are useless if urls are specified. Use the special form "trending:genre:time" instead.

This also includes a lot of refactoring and smaller improvements.